Base_dados/requirements.txt
torch
ultralytics
opencv-python
numpy
matplotlib
PyYAML
tqdm